+++ tests/integration/flows/test_duckdb_to_sql_server.py
@@ -1,17 +1,16 @@
 import json
-import logging
-import os
-from unittest import mock
-
 import pytest
+import os
+import logging
 from prefect.tasks.secrets import PrefectSecret
 
+from viadot.tasks.azure_key_vault import AzureKeyVaultSecret
 from viadot.flows import DuckDBToSQLServer
 from viadot.sources import DuckDB
-from viadot.tasks.azure_key_vault import AzureKeyVaultSecret
 
 TABLE = "test_table"
-SCHEMA = "test_schema"
+DUCKDB_SCHEMA = "test_schema"
+SQL_SERVER_SCHEMA = "sandbox"
 TABLE_MULTIPLE_PARQUETS = "test_multiple_parquets"
 DATABASE_PATH = "test_db_123.duckdb"
 
@@ -23,24 +22,13 @@ def duckdb():
     os.remove(DATABASE_PATH)
 
 
-def test__check_if_schema_exists(duckdb):
-    duckdb.run(f"CREATE SCHEMA {SCHEMA}")
-    assert not duckdb._check_if_schema_exists(SCHEMA)
-
-
 def test_create_table_from_parquet(duckdb, TEST_PARQUET_FILE_PATH, caplog):
     with caplog.at_level(logging.INFO):
         duckdb.create_table_from_parquet(
-            schema=SCHEMA, table=TABLE, path=TEST_PARQUET_FILE_PATH
+            schema=DUCKDB_SCHEMA, table=TABLE, path=TEST_PARQUET_FILE_PATH
         )
 
     assert "created successfully" in caplog.text
-
-
-def test_duckdb_sql_server_init():
-
-    flow = DuckDBToSQLServer("test_duckdb_init")
-    assert flow
 
 
 def test_duckdb_sql_server_flow():
@@ -60,20 +48,8 @@ def test_duckdb_sql_server_flow():
         sql_server_credentials=json.loads(credentials_str),
         sql_server_schema="sandbox",
         sql_server_table=TABLE,
-        duckdb_schema=SCHEMA,
+        duckdb_schema=DUCKDB_SCHEMA,
         duckdb_table=TABLE,
     )
     r = flow.run()
     assert r.is_successful()
-
-
-def test_duckdb_sql_server_flow_mocked():
-    with mock.patch.object(DuckDBToSQLServer, "run", return_value=True) as mock_method:
-        flow = DuckDBToSQLServer(
-            "test_duckdb_flow_run",
-            sql_server_table=TABLE,
-            duckdb_schema=SCHEMA,
-            duckdb_table=TABLE,
-        )
-        flow.run()
-        mock_method.assert_called_with()