+++ viadot/sources/__init__.py
@@ -7,6 +7,9 @@ from .outlook import Outlook
 from .salesforce import Salesforce
 from .sharepoint import Sharepoint
 from .supermetrics import Supermetrics
+from .genesys import Genesys
+from .sftp import SftpConnector
+
 
 try:
     from .sap_rfc import SAPRFC